101-mul.c

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@
  */
 int main(int argc, char **argv)
 {
-	int a, b;
+	long int a, b;
 
 	if (argc != 3)
 	{
@@ -18,6 +18,6 @@ int main(int argc, char **argv)
 	}
 	a = atoi(*(argv + 1));
 	b = atoi(*(argv + 2));
-	printf("%d\n", a * b);
+	printf("%ld\n", a * b);
 	return (0);
 }
